# [AOT] Enable AOT for CmdPal