--- a/.pipelines/v2/release.yml
+++ b/.pipelines/v2/release.yml
@@ -38,6 +38,11 @@ parameters:
     displayName: "Build Using Visual Studio Preview"
     default: false
 
+  - name: enableAOT
+    type: boolean
+    displayName: "Enable AOT (Ahead-of-Time) Compilation for CmdPal"
+    default: true
+
 name: $(BuildDefinitionName)_$(date:yyMM).$(date:dd)$(rev:rrr)
 
 variables:
@@ -95,7 +100,7 @@ extends:
                 useManagedIdentity: $(SigningUseManagedIdentity)
                 clientId: $(SigningOriginalClientId)
               # Have msbuild use the release nuget config profile
-              additionalBuildOptions: /p:RestoreConfigFile="$(Build.SourcesDirectory)\.pipelines\release-nuget.config"
+              additionalBuildOptions: /p:RestoreConfigFile="$(Build.SourcesDirectory)\.pipelines\release-nuget.config" /p:EnableCmdPalAOT=${{ parameters.enableAOT }}
               beforeBuildSteps:
                 # Sets versions for all PowerToy created DLLs
                 - pwsh: |-