--- a/NOTICE.md
+++ b/NOTICE.md
@@ -1580,4 +1580,3 @@ SOFTWARE.
 - WinUIEx 2.2.0
 - WPF-UI 3.0.5
 - WyHash 1.0.5
-