--- a/nuget.config
+++ b/nuget.config
@@ -2,10 +2,10 @@
 <configuration>
 	<packageSources>
 		<clear />
-		<add key="nuget.org" value="https://api.nuget.org/v3/index.json" />
+		<add key="PowerToysPublicDependencies" value="https://pkgs.dev.azure.com/shine-oss/PowerToys/_packaging/PowerToysPublicDependencies/nuget/v3/index.json" />
 	</packageSources>
 	<packageSourceMapping>
-		<packageSource key="nuget.org">
+		<packageSource key="PowerToysPublicDependencies">
 			<package pattern="*" />
 		</packageSource>
 	</packageSourceMapping>