--- a/src/modules/cmdpal/Microsoft.CmdPal.UI/ExtViews/ContentPage.xaml
+++ b/src/modules/cmdpal/Microsoft.CmdPal.UI/ExtViews/ContentPage.xaml
@@ -10,7 +10,9 @@
     xmlns:controls="using:CommunityToolkit.WinUI.Controls"
     xmlns:converters="using:CommunityToolkit.WinUI.Converters"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:labToolkit="using:CommunityToolkit.Labs.WinUI.MarkdownTextBlock"
     xmlns:local="using:Microsoft.CmdPal.UI"
+    xmlns:markdownTextBlockRns="using:CommunityToolkit.WinUI.Controls.MarkdownTextBlockRns"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:toolkit="using:CommunityToolkit.WinUI.UI.Controls"
     xmlns:viewModels="using:Microsoft.CmdPal.UI.ViewModels"
@@ -19,6 +21,12 @@
 
     <Page.Resources>
         <ResourceDictionary>
+            <markdownTextBlockRns:MarkdownThemes
+                x:Key="DefaultMarkdownThemeConfig"
+                H3FontSize="12"
+                H3FontWeight="Normal" />
+            <labToolkit:MarkdownConfig x:Key="DefaultMarkdownConfig" Themes="{StaticResource DefaultMarkdownThemeConfig}" />
+
             <StackLayout
                 x:Name="VerticalStackLayout"
                 Orientation="Vertical"
@@ -43,12 +51,9 @@
 
             <DataTemplate x:Key="MarkdownContentTemplate" x:DataType="viewModels:ContentMarkdownViewModel">
                 <Grid Margin="0,4,4,4" Padding="12,8,8,8">
-                    <toolkit:MarkdownTextBlock
+                    <labToolkit:MarkdownTextBlock
                         Background="Transparent"
-                        Header3FontSize="12"
-                        Header3FontWeight="Normal"
-                        Header3Foreground="{ThemeResource TextFillColorSecondaryBrush}"
-                        IsTextSelectionEnabled="True"
+                        Config="{StaticResource DefaultMarkdownConfig}"
                         Text="{x:Bind Body, Mode=OneWay}" />
                 </Grid>
             </DataTemplate>
@@ -61,12 +66,9 @@
 
             <DataTemplate x:Key="NestedMarkdownContentTemplate" x:DataType="viewModels:ContentMarkdownViewModel">
                 <Grid>
-                    <toolkit:MarkdownTextBlock
+                    <labToolkit:MarkdownTextBlock
                         Background="Transparent"
-                        Header3FontSize="12"
-                        Header3FontWeight="Normal"
-                        Header3Foreground="{ThemeResource TextFillColorSecondaryBrush}"
-                        IsTextSelectionEnabled="True"
+                        Config="{StaticResource DefaultMarkdownConfig}"
                         Text="{x:Bind Body, Mode=OneWay}" />
                 </Grid>
             </DataTemplate>