@crutkas 2025-07-11
I’m very interested in this from size to speed