@moooyo 2025-07-11
> I’m very interested in this from size to speed

I'm trying to do a basic test.