@moooyo 2025-07-14
> # AOT Performance Improvement Summary
> By enabling AOT in cmdpal, we successfully achieved:
> 
> * **56.4% reduction** in installed size
> * **3.5x performance improvement** in built-in extension loading logic
> * **1.6x performance improvement** in overall page loading
> 
> ## Key Technical Indicators
> Branch	Installed Size (MB)	Memory Spike (MB)	Memory Usage (MB)	Main Window Rendering Complete (ms)	Built-in Extension Loaded Complete (ms)	Main Page Loading Complete (ms)	Built-in Extension Loading Time (ms)
> 91.1	121	196.84	94.76	Not measured	1469.20	1502.80	Not measured
> 92.1	121	192.88	96.80	Not measured	1426.00	1457.40	Not measured
> Main branch (disable AOT)	123	197.74	95.36	899.0	1434	1467	846.8
> Main branch (enable AOT)	53.6	160.56	95.10	809.60	895.80	909.80	246.8
> ## Raw Data
> For detailed information on how this data was collected, please refer to branch: `yuleng/aot/benchmark/93-aot`
> 
> All date listed below is unix ms timestamp.
> 
> ### Version 0.91.1
> Memory Spike (MB)	Memory Stable (MB)	OnLaunched	ShowHwnd Complete	LoadBuiltinsAsync Complete	NavigateToPageMessage
> 199.7	90.3	1752480277819	-	1752480279551	1752480279585
> 199.2	88.7	1752480393206	-	1752480394735	1752480394766
> 199.6	89.7	1752480471467	-	1752480472766	1752480472798
> 192.2	101.9	1752480520061	-	1752480521499	1752480521539
> 193.5	103.2	1752480562058	-	1752480563406	1752480563437
> ### Version 0.92.1
> Memory Spike (MB)	Memory Stable (MB)	OnLaunched	ShowHwnd Complete	LoadBuiltinsAsync Complete	NavigateToPageMessage
> 202.8	89.3	1752480848898	-	1752480850322	1752480850352
> 179	102.6	1752480918069	-	1752480919833	1752480919864
> 203.5	88.2	1752480969369	-	1752480970693	1752480970725
> 204.2	101.6	1752481014726	-	1752481016014	1752481016046
> 174.9	102.3	1752481057186	-	1752481058516	1752481058548
> ### Main Branch (AOT Disabled)
> Memory Spike (MB)	Memory Stable (MB)	OnLaunched	ShowHwnd Complete	LoadBuiltinsAsync Complete	NavigateToPageMessage	Built-in Extension Loading Time (ms)
> 208.3	90.7	1752482085050	1752482085861	1752482086461	1752482086493	862
> 202.1	90.1	1752482141694	1752482142589	1752482143080	1752482143112	789
> 199.2	88.3	1752482195196	1752482196223	1752482196786	1752482196819	973
> 198.7	103.9	1752482252775	1752482253632	1752482254104	1752482254133	746
> 180.4	103.8	1752482304509	1752482305414	1752482305965	1752482306002	864
> ### Main Branch (AOT Enabled)
> Memory Spike (MB)	Memory Stable (MB)	OnLaunched	ShowHwnd Complete	LoadBuiltinsAsync Complete	NavigateToPageMessage	Built-in Extension Loading Time (ms)
> 163.5	95.3	1752483234513	1752483235309	1752483235398	1752483235411	242
> 140.6	92.9	1752483291364	1752483292182	1752483292267	1752483292279	264
> 170.6	94.3	1752483362794	1752483363625	1752483363725	1752483363740	278
> 171.0	95.7	1752483413076	1752483413892	1752483413970	1752483413984	217
> 157.1	97.3	1752483462323	1752483463110	1752483463189	1752483463205	233

@zadjii-msft @crutkas 

I forgot to add some metrics in the previous version to measure the performance so we can get an overall data to analysis.

currently I can only measure it in my devbox...